+++ game/game/g_team.cpp
@@ -1105,7 +1105,7 @@ void TeamplayInfoMessage( gentity_t *ent ) {
 //				level.sortedClients[i], player->client->pers.teamState.location, h, a, 
 				i, player->client->pers.teamState.location, h, a, 
 				player->client->ps.weapon, player->s.powerups);
-			j = strlen(entry);
+			j = (int)( strlen(entry) );
 			if ((size_t)( stringlength + j ) > sizeof(string))
 				break;
 			strcpy (string + stringlength, entry);
